+++ CHANGELOG.md
@@ -9,12 +9,13 @@
 
 ## 0.1.2
 
-Released on ??
+Released on 13/12/2020
 
 - General performance and code improvements
 - Improved symlinks management
 - Keybindings:
   - `E`: Delete file (Same as `DEL`); added because some keyboards don't have `DEL` (hey, that's my MacBook Air's keyboard!)
+  - `Ctrl+C`: Abort transfer process
 
 ## 0.1.1